Implemented Parser
In order to provide string splitting capability this commit introduces a
module called Parser in the lib directory. It takes two arguments,
number and chunk. We intentionally kept this functionality away from
Rails so that we can test in isolation, then implement the feature
within Rails knowing our code is solid.