Implement the HTTP End Point
This commit introduces the new API end point /commafy which utilises the
parser we implemented earlier. The implementation uses the Grape API gem
which enables the code to remain "change friendly" in the future. The
end point is tested and development of this feature highlighted the need
for an additional method in our library to convert a string to either an
integer or a float appropriately.